+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2021-06-09
+### Added
+- Smeagol user agent provider ([#6](https://github.com/scm-manager/scm-smeagol-plugin/pull/6))
+
+## 1.1.2 - 2021-04-27
+### Fixed
+- Synchronization of computation of smeagol information ([#5](https://github.com/scm-manager/scm-smeagol-plugin/pull/5))
+
 ## 1.1.1 - 2021-04-07
 ### Fixed
 - Repository overview which contains subversion repositories ([#4](https://github.com/scm-manager/scm-smeagol-plugin/pull/4))